+++ src/AI/cubeMonster/cubeMonster.cpp
@@ -1,0 +1,13 @@
+#include "cubeMonster.hpp"
+
+namespace CE {
+	namespace AI {
+		CubeMonster::CubeMonster() {
+			//
+		}
+
+		CubeMonster::~CubeMonster() {
+			//
+		}
+	}
+}